% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{construct_adj_long}
\alias{construct_adj_long}
\title{The distance matrix undirected network was constructed to calculate the distance between any two spots}
\usage{
construct_adj_long(M)
}
\arguments{
\item{M}{A data frame containing the coordinates of the idle data and the barcode}
}
\value{
A data frame containing the Euclidean distance between any two spots
}
\description{
The distance matrix undirected network was constructed to calculate the distance between any two spots
}
\examples{
\dontrun{
dist_long <- construct_adj_long(meta)
}
}
