% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{construct_adj_wide}
\alias{construct_adj_wide}
\title{Construct the weighted adjacency matrix}
\usage{
construct_adj_wide(dist_long_sub)
}
\arguments{
\item{dist_long_sub:}{A data frame including the distance between any two spots}
}
\description{
Construct the weighted adjacency matrix
}
\examples{
Network <- construct_adj_wide(dist_long)


}
