% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{hierarchy}
\alias{hierarchy}
\title{Define hierarchy}
\usage{
hierarchy(plot_all)
}
\arguments{
\item{plot_all}{Results for the scores as well as p-values}
}
\value{
A data frame containing hierarchical information
}
\description{
Define hierarchy
}
\examples{
plot_all <- hierarchy(plot_all)


}
