% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{high_ppi_list}
\alias{high_ppi_list}
\title{High confidence PPI list}
\format{
A list where each element is a data frame with columns `source` and `target`.
}
\source{
Processed from high-confidence.csv
}
\usage{
high_ppi_list
}
\description{
A list of protein-protein interactions for high confidence pathways.
}
\keyword{datasets}
