% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{hot_region_plot}
\alias{hot_region_plot}
\title{Output region results}
\usage{
hot_region_plot(out_data, pathway)
}
\arguments{
\item{out_data}{output of 'compute_siossar'}

\item{pathway}{pathway}
}
\value{
Output graph
}
\description{
Output region results
}
\examples{
\dontrun{
p <- hot_region_plot(object, "glycolytic_process")
}
}
