% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{hplot}
\alias{hplot}
\title{Output hierarchical results}
\usage{
hplot(out_data, pathway, show_layers = 1:10, layer_colors = NULL)
}
\arguments{
\item{out_data}{output of 'compute_siossar'}

\item{pathway}{pathway}

\item{show_layers}{numeric vector of layers to display,Default is 10}

\item{layer_colors}{optional named color vector}
}
\value{
Output graph
}
\description{
Output hierarchical results
}
\examples{
\dontrun{
p <- hplot(object, "glycolytic_process",1:3)
}
}
