% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{ISs}
\alias{ISs}
\title{Integrated scores}
\usage{
ISs(aa, topn_sm, SM_exp)
}
\arguments{
\item{aa}{A data frame, run_rwr function run results}

\item{topn_sm}{A data frame containing information on the mean expression of secreted proteins used as seed spots}

\item{SM_exp}{A data frame containing information on the average secreted protein expression for each spot and whether the expression content is zero}
}
\description{
Integrated scores
}
\examples{
\dontrun{
# These objects are typically generated by the SIOSSAR workflow.
# Example :
# aa <- run_rwr(Network, all_spot, topn_sm)
# topn_sm <- top.secret(SM_exp)
# SM_exp <- secret.mean(exp, genes)
IS <- ISs(aa, topn_sm, SM_exp)
}


}
