% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{low_ppi_list}
\alias{low_ppi_list}
\title{Low confidence PPI list}
\format{
A list where each element is a data frame with columns `source` and `target`.
}
\source{
Processed from low-confidence.csv
}
\usage{
low_ppi_list
}
\description{
A list of protein-protein interactions for low confidence pathways.
}
\keyword{datasets}
