% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medium_ppi_list}
\alias{medium_ppi_list}
\title{Medium confidence PPI list}
\format{
A list where each element is a data frame with columns `source` and `target`.
}
\source{
Processed from medium-confidence.csv
}
\usage{
medium_ppi_list
}
\description{
A list of protein-protein interactions for medium confidence pathways.
}
\keyword{datasets}
