% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{membrane_mean}
\alias{membrane_mean}
\title{Membrane protein expression profile}
\usage{
membrane_mean(exp, gene)
}
\arguments{
\item{exp}{Expression profiling data with genes as rows and genomic loci as columns}

\item{gene}{A vector, a membrane protein gene}
}
\description{
Membrane protein expression profile
}
\examples{
\dontrun{
mm <- membrane_mean(exp, gene = c("CCR7", "SRC"))
}

}
