% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{norm}
\alias{norm}
\title{Normalization of expression profiles}
\usage{
norm(object, norm = "NormalizeData")
}
\arguments{
\item{object:}{a seurat object}

\item{norm:}{normalization method,NormalizeData or SCTransform}
}
\value{
returns a dataframe with genes as rows and genomic loci as columns
}
\description{
Normalization of expression profiles
}
\examples{
exp <- norm(object)


}
