% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{plot_PPI_sankey}
\alias{plot_PPI_sankey}
\title{Output Contribution of SP-MP}
\usage{
plot_PPI_sankey(object, pathway)
}
\arguments{
\item{object}{output of 'compute_siossar'}

\item{pathway}{pathway}
}
\value{
Output graph
}
\description{
Output Contribution of SP-MP
}
\examples{
\dontrun{
p <- plot_PPI_sankey(object,"glycolytic_process")
}

}
