% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{plot_top_pathway_heatmap}
\alias{plot_top_pathway_heatmap}
\title{Output top 20 pathway}
\usage{
plot_top_pathway_heatmap(object, pathway, top_n = 20)
}
\arguments{
\item{object}{output of 'compute_siossar'}

\item{pathway}{pathway}

\item{top_n}{Select the number of top pathways to display based on the count of activated spots}
}
\value{
Output graph
}
\description{
Output top 20 pathway
}
\examples{
\dontrun{
p <- plot_top_pathway_heatmap(object, pathway, top_n = 1)
}
}
