% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{RandomWalk2igraph}
\alias{RandomWalk2igraph}
\title{RandomWalk function}
\usage{
RandomWalk2igraph(igraphM, VertexWeight, EdgeWeight = TRUE, gamma = 0.7)
}
\arguments{
\item{igraphM}{matrix of adjacency,the rows and columns are spot}

\item{VertexWeight}{a vector v with a length equal to the total number of spots in the slice}

\item{EdgeWeight}{logical value, default is TRUE}

\item{gamma}{restart probability,default is 0.7}
}
\value{
returns a vector ,reflecting the probability each spot is labeled during the propagation
}
\description{
RandomWalk function
}
\examples{
\dontrun{
# This example requires a valid adjacency matrix and seed vector
# visProbs <- RandomWalk2igraph(matrix1, Vector1, EdgeWeight = TRUE, gamma = 0.7)
}

}
