% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{score_plot}
\alias{score_plot}
\title{Output score results}
\usage{
score_plot(out_data, pathway)
}
\arguments{
\item{out_data}{output of 'compute_siossar'}

\item{pathway}{pathway}
}
\value{
Output graph
}
\description{
Output score results
}
\examples{
p <- score_plot(object,"glycolytic_process")


}
