% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{secret_mean}
\alias{secret_mean}
\title{Expression profile of secreted proteins}
\usage{
secret_mean(exp, gene)
}
\arguments{
\item{exp}{Expression profiling data with genes as rows and genomic loci as columns}

\item{gene}{A vector, a secreted protein gene}
}
\description{
Expression profile of secreted proteins
}
\examples{
\dontrun{
sm <- secret_mean(exp,gene=c("CCL21","ITGB1"))
}

}
