% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIOSSAR_functions.R
\name{top_secret}
\alias{top_secret}
\title{Filter the number of seed nodes}
\usage{
top_secret(SM_exp)
}
\arguments{
\item{SM_exp}{A data frame containing information on the average secreted protein expression for each spot and whether the expression content is zero}
}
\description{
Filter the number of seed nodes
}
\examples{
\dontrun{
topn_sm <- top_secret(SM_exp)
}

}
